+++ lib/mangoapps/modules/recognitions.rb
@@ -5,6 +5,9 @@ require_relative "recognitions/core_value_tags"
 require_relative "recognitions/leaderboard_info"
 require_relative "recognitions/tango_gift_cards"
 require_relative "recognitions/gift_cards"
+require_relative "recognitions/get_awards_list"
+require_relative "recognitions/get_profile_awards"
+require_relative "recognitions/get_team_awards"
 
 module MangoApps
   class Client
@@ -15,6 +18,9 @@ module MangoApps
       include MangoApps::Client::Recognitions::LeaderboardInfo
       include MangoApps::Client::Recognitions::TangoGiftCards
       include MangoApps::Client::Recognitions::GiftCards
+      include MangoApps::Client::Recognitions::GetAwardsList
+      include MangoApps::Client::Recognitions::GetProfileAwards
+      include MangoApps::Client::Recognitions::GetTeamAwards
     end
   end
 end